Introduce completion test utils
